restructured project into cli and core
