+++ Cargo.toml
@@ -1,5 +1,8 @@
-[package]
-name = "epimetheus"
+[workspace]
+members = ["methylome", "epimetheus-core", "epimetheus-cli"]
+resolver = "2"
+
+[workspace.package]
 version = "0.5.2"
 edition = "2021"
 authors = ["semoda@bio.aau.dk"]
@@ -8,27 +11,14 @@ description = "Cli tool for fast lookup in pileup for motif methylation"
 license = "MIT" 
 repository = "https://github.com/SebastianDall/epimetheus"
 
-[workspace]
-members = ["methylome", "."]
 
-[dependencies]
+[workspace.dependencies]
 clap = { version = "4.5.20", features = ["derive"] }
-seq_io = "0.3.2"
-methylome = { path = "methylome"}
-rayon = "1.10.0"
-indicatif = {version = "0.17.8", features = ["rayon"]}
+anyhow = "1.0.94"
 log = "0.4.22"
 env_logger = "0.11.5"
 humantime = "2.1.0"
-anyhow = "1.0.94"
-tempfile = "3.14.0"
-bytesize = "1.3.0"
-csv = "1.3.1"
-ahash = "0.8.11"
-# pprof = { version = "0.14", features = ["flamegraph"] }
 
-[dev-dependencies]
-criterion = "0.5.1"
 
 [profile.release]
 debug = true
