+++ epimetheus-core/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod data;
+pub mod data_load;
+pub mod extract_methylation_pattern;
+pub mod motif_clustering;
+pub mod processing;
+pub mod utils;
+
+pub use extract_methylation_pattern::extract_methylation_pattern;
+pub use motif_clustering::motif_clustering;
