--- tests/data/expected_out.tsv
@@ -1,7 +1,0 @@
-contig	motif	mod_type	mod_position	median	mean_read_cov	N_motif_obs	motif_occurences_total
-contig_2	GATC	a	1	0.847780487804878	130.90394736842106	760	760
-contig_2	GATC	m	3	0	126.26253298153034	758	760
-contig_2	RGATCY	a	2	0.7877252252252251	126.40384615384616	52	52
-contig_3	GATC	a	1	0.8940499889404998	310.0593220338983	708	708
-contig_3	GATC	m	3	0.0031645569620253164	298.04802259887003	708	708
-contig_3	RGATCY	a	2	0.8530862923203963	301.2142857142857	84	84
